# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Steam4Caster API",
    "description": "\nBackend API for **Steam4Caster**: probabilistic Steam discount forecasts, explainable\nBUY / WAIT / NEUTRAL recommendations, watchlists and alerts.\n\n* All public routes are versioned under `/api/v1`.\n* Authenticate with `Authorization: Bearer <access token>`.\n* Money is always integer minor units plus an ISO 4217 currency, per region. Prices are\n  never converted between currencies.\n* Errors use one envelope: `{\"error\": {\"code\", \"message\", \"correlation_id\", \"details\"}}`.\n* Forecasts and recommendations are statistical estimates based on past prices. They are not guarantees and not financial advice.\n\nPrice data is provided by the IsThereAnyDeal API. Steam4Caster is not affiliated with\nIsThereAnyDeal or Valve.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/watchlist": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "List the user's watchlist",
        "operationId": "list_watchlist_api_v1_watchlist_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistEntryOut"
                  },
                  "type": "array",
                  "title": "Response List Watchlist Api V1 Watchlist Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "watchlist"
        ],
        "summary": "Add a game to the watchlist",
        "operationId": "create_entry_api_v1_watchlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/watchlist/summary": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "Portfolio-style totals and recommendation counts",
        "description": "Current cost, cost at recorded historical lows and modelled expected cost per horizon, grouped by currency. Modelled values are estimates.",
        "operationId": "summary_api_v1_watchlist_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistSummary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/watchlist/{entry_id}": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "Get one watchlist entry",
        "operationId": "get_entry_api_v1_watchlist__entry_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "patch": {
        "tags": [
          "watchlist"
        ],
        "summary": "Update alert preferences for an entry",
        "description": "Only fields present in the body are changed. Send an explicit null to clear target_price_minor, min_discount_pct or max_wait_days.",
        "operationId": "update_entry_api_v1_watchlist__entry_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "watchlist"
        ],
        "summary": "Remove a watchlist entry",
        "operationId": "delete_entry_api_v1_watchlist__entry_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntryPrice": {
        "properties": {
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "regular": {
            "$ref": "#/components/schemas/Money"
          },
          "discount_pct": {
            "type": "integer",
            "title": "Discount Pct"
          },
          "historical_low": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "price",
          "regular",
          "discount_pct",
          "historical_low",
          "observed_at",
          "url"
        ],
        "title": "EntryPrice"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "GameOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "itad_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Itad Id",
            "description": "IsThereAnyDeal game id."
          },
          "steam_app_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steam App Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "mature": {
            "type": "boolean",
            "title": "Mature"
          },
          "early_access": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Early Access"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "developers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Developers"
          },
          "publishers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Publishers"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "assets": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Assets"
          },
          "provider_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "itad_id",
          "steam_app_id",
          "title",
          "slug",
          "type",
          "mature",
          "early_access",
          "release_date",
          "developers",
          "publishers",
          "tags",
          "assets",
          "provider_url"
        ],
        "title": "GameOut"
      },
...
      "Money": {
        "properties": {
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor",
            "examples": [
              149900
            ]
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "examples": [
              "INR"
            ]
          },
          "amount": {
            "type": "string",
            "title": "Amount",
            "examples": [
              "1499.00"
            ]
          }
        },
        "type": "object",
        "required": [
          "amount_minor",
          "currency",
          "amount"
        ],
        "title": "Money",
        "description": "An amount in integer minor units with its ISO 4217 currency. ``amount`` is a\ndecimal string for display; clients must not parse it into binary floats."
      },
...
      "WatchlistCreate": {
        "properties": {
          "game_id": {
            "type": "string",
            "format": "uuid",
            "title": "Game Id"
          },
          "country": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country",
            "description": "Defaults to the user's country."
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency",
            "description": "Currency of target_price_minor; must match the region's."
          },
          "target_price_minor": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Price Minor"
          },
          "min_discount_pct": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Discount Pct"
          },
          "max_wait_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 365.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Wait Days"
          },
          "historical_low_only": {
            "type": "boolean",
            "title": "Historical Low Only",
            "default": false
          },
          "notify_on_buy": {
            "type": "boolean",
            "title": "Notify On Buy",
            "default": true
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Channel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          }
        },
        "type": "object",
        "required": [
          "game_id"
        ],
        "title": "WatchlistCreate",
        "example": {
          "channels": [
            "WEB_PUSH",
            "EMAIL"
          ],
          "country": "IN",
          "currency": "INR",
          "game_id": "0b0f6c3e-7d0e-4f7e-9d43-0e9d1f0c2a11",
          "historical_low_only": false,
          "max_wait_days": 60,
          "min_discount_pct": 50,
          "notify_on_buy": true,
          "target_price_minor": 74900
        }
      },
      "WatchlistEntryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "game": {
            "$ref": "#/components/schemas/GameOut"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "target_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "min_discount_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Discount Pct"
          },
          "max_wait_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Wait Days"
          },
          "historical_low_only": {
            "type": "boolean",
            "title": "Historical Low Only"
          },
          "notify_on_buy": {
            "type": "boolean",
            "title": "Notify On Buy"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "current": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntryPrice"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null until a regional price is known."
          },
          "forecast": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntryForecast"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "game",
          "country",
          "currency",
          "target_price",
          "min_discount_pct",
          "max_wait_days",
          "historical_low_only",
          "notify_on_buy",
          "channels",
          "is_active",
          "created_at",
          "updated_at",
          "current",
          "forecast",
          "recommendation"
        ],
        "title": "WatchlistEntryOut"
      },
      "WatchlistSummary": {
        "properties": {
          "entries": {
            "type": "integer",
            "title": "Entries"
          },
          "totals": {
            "items": {
              "$ref": "#/components/schemas/CurrencyTotals"
            },
            "type": "array",
            "title": "Totals",
            "description": "One block per currency. Amounts are never summed across currencies."
          },
          "recommendation_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Recommendation Counts"
          },
          "likely_on_sale_within_30d": {
            "items": {
              "$ref": "#/components/schemas/LikelySale"
            },
            "type": "array",
            "title": "Likely On Sale Within 30D"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Forecasts and recommendations are statistical estimates based on past prices. They are not guarantees and not financial advice."
          }
        },
        "type": "object",
        "required": [
          "entries",
          "totals",
          "recommendation_counts",
          "likely_on_sale_within_30d"
        ],
        "title": "WatchlistSummary"
      },
      "WatchlistUpdate": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "target_price_minor": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Price Minor"
          },
          "min_discount_pct": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Discount Pct"
          },
          "max_wait_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 365.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Wait Days"
          },
          "historical_low_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Historical Low Only"
          },
          "notify_on_buy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notify On Buy"
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Channel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "WatchlistUpdate"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /api/v1/auth/login.",
        "scheme": "bearer"
      }
    }
  }
}